requirements.txt
streamlit>=1.39.0
google-genai>=1.0.0
chromadb>=0.5.0
PyMuPDF>=1.24.0
python-dotenv>=1.0.1
pytest>=8.0.0
